Display the progress of desktop app installation before starting the desktop
This just prevents people from seeing a black screen for a long time and assuming something is going wrong.
This initial setup is installaing flatpaks to the mounted home directory. This will not need to happen on every container startup. So this black screen is only a problem on the first start. But this change at least gives people something to look at.

Closes #50