Improvements to NVIDIA Patch installation
The GitHub servers had a small outage today while I was working. This prevented the containers from being able to start.
This change allows for the patch to be re-used between container re-creations and also prevents the failure from stopping boot. If the patch fails to download or install, then it will continue with an unpatched driver.